more'

+++ flake/process-compose.nix
@@ -11,14 +11,18 @@
             ${self.inputs.cosmos.packages.${system}.osmosis}/bin/osmosisd "$@"
           '';
         };
+        hermes = self.inputs.cosmos.packages.${system}.hermes;
+
         osmosisd-gen = pkgs.writeShellApplication {
           name = "osmosisd-gen";
           runtimeInputs = [ osmosisd pkgs.jq pkgs.yq pkgs.dasel ];
           text = ''
-            HOME=${devnet-root-directory}
-            export HOME
-            OSMOSIS_DATA="$HOME/.osmosisd"
-            rm --force --recursive "$OSMOSIS_DATA" 
+            OSMOSIS_DATA="${devnet-root-directory}/.osmosisd"
+            REUSE=true
+            export REUSE
+            if [[ $REUSE == false ]]; then
+              rm --force --recursive "$OSMOSIS_DATA" 
+            fi
             CHAIN_ID="osmosis"
             VALIDATOR_MONIKER="validator"
             VALIDATOR_MNEMONIC="bottom loan skill merry east cradle onion journey palm apology verb edit desert impose absurd oil bubble sweet glove shallow size build burst effort"
@@ -26,7 +30,7 @@
             RELAYER_MNEMONIC="black frequent sponsor nice claim rally hunt suit parent size stumble expire forest avocado mistake agree trend witness lounge shiver image smoke stool chicken"
             CONFIG_FOLDER=$OSMOSIS_DATA/config
             GENESIS=$CONFIG_FOLDER/genesis.json
-            mkdir --parents $OSMOSIS_DATA
+            mkdir --parents "$OSMOSIS_DATA/data/cs.wal"
 
             echo "$VALIDATOR_MNEMONIC" | osmosisd init --chain-id="$CHAIN_ID" --home "$OSMOSIS_DATA" --recover "$VALIDATOR_MONIKER"
 
@@ -102,7 +106,7 @@
             dasel put --type string --file "$CONFIG_FOLDER/config.toml" --value "0.0.0.0:16060" '.rpc.pprof_laddr'
             dasel put --type string --file "$CONFIG_FOLDER/config.toml" --value "tcp://127.0.0.1:36658" '.proxy_app'
             dasel put --type string --file "$CONFIG_FOLDER/config.toml" --value ":36660" '.instrumentation.prometheus_listen_addr'
-            dasel put --type string --file "$CONFIG_FOLDER/config.toml" --value "tcp://0.0.0.0:36656" '.p2p.prometheus_listen_addr'
+            dasel put --type string --file "$CONFIG_FOLDER/config.toml" --value "tcp://0.0.0.0:36656" '.p2p.laddr'
             dasel put --type string --file "$CONFIG_FOLDER/config.toml" --value "tcp://localhost:36657" '.node'
 
             dasel put --type string --file "$CONFIG_FOLDER/app.toml" --value "0.0.0.0:19090" '.grpc.address'
@@ -119,7 +123,13 @@
             dasel put --type bool --file $CONFIG_FOLDER/app.toml --value "true" '.api.enabled-unsafe-cors'
             dasel put --type bool --file $CONFIG_FOLDER/app.toml --value "true" '.grpc-web.enable-unsafe-cors'
 
-            osmosisd start --home "$OSMOSIS_DATA"
+            dasel put --type string --file $CONFIG_FOLDER/client.toml --value "tcp://localhost:36657" '.node'
+            dasel put --type string --file $CONFIG_FOLDER/client.toml --value "$CHAIN_ID" '.chain-id'
+            dasel put --type string --file $CONFIG_FOLDER/client.toml --value "test" '.keyring-backend'
+            dasel put --type string --file $CONFIG_FOLDER/client.toml --value "json" '.output'
+
+
+            osmosisd start --home "$OSMOSIS_DATA" --rpc.unsafe --rpc.laddr tcp://0.0.0.0:36657 --pruning=nothing --grpc.address localhost:19090   --address "tcp://0.0.0.0:36658" --p2p.external-address 43421 --p2p.laddr "tcp://0.0.0.0:36656" --p2p.pex false --p2p.upnp  false  --p2p.seed_mode true
           '';
         };
 
@@ -141,27 +151,30 @@
       process-compose.devnet-xc = {
         settings = {
           processes = {
-            centauri = {
-              command = self'.packages.centaurid-gen;
-              readiness_probe.http_get = {
-                host = "127.0.0.1";
-                port = 26657;
-              };
-            };
+            # centauri = {
+            #   command = self'.packages.centaurid-gen;
+            #   readiness_probe.http_get = {
+            #     host = "127.0.0.1";
+            #     port = 26657;
+            #   };
+            # };
+            # centauri-init = {
+            #   command = self'.packages.centaurid-init;
+            #   depends_on."centauri".condition = "process_healthy";
+            #   log_location = "/tmp/composable-devnet/centauri-init.log";
+            # };
             osmosis = {
               command = self'.packages.osmosisd-gen;
               readiness_probe.http_get = {
                 host = "127.0.0.1";
                 port = 36657;
               };
+              log_location = "/tmp/composable-devnet/osmosis.log";
             };
             osmosis-init = {
               command = self'.packages.osmosisd-init;
               depends_on."osmosis".condition = "process_healthy";
-            };
-            centauri-init = {
-              command = self'.packages.centaurid-init;
-              depends_on."centauri".condition = "process_healthy";
+              log_location = "/tmp/composable-devnet/osmosis-init.log";
             };
             picasso = {
               command = self'.packages.zombienet-rococo-local-picasso-dev;
@@ -169,75 +182,75 @@
               log_location = "/tmp/composable-devnet/zombienet.log";
             };
 
-            hyperspace-client = {
-              command = ''
-                sleep 20
-                COMPOSABLE_DATA=/tmp/composable-devnet/
-                HYPERSPACE_DATA="$COMPOSABLE_DATA/hyperspace"
-                RUST_LOG="hyperspace=trace,hyperspace_parachain=trace,hyperspace_cosmos=trace"
-                export RUST_LOG
-                mkdir --parents "$COMPOSABLE_DATA"
-                mkdir --parents "$HYPERSPACE_DATA"
+            # hyperspace-client = {
+            #   command = ''
+            #     sleep 20
+            #     COMPOSABLE_DATA=/tmp/composable-devnet/
+            #     HYPERSPACE_DATA="$COMPOSABLE_DATA/hyperspace"
+            #     RUST_LOG="hyperspace=trace,hyperspace_parachain=trace,hyperspace_cosmos=trace"
+            #     export RUST_LOG
+            #     mkdir --parents "$COMPOSABLE_DATA"
+            #     mkdir --parents "$HYPERSPACE_DATA"
 
-                cp --dereference --no-preserve=mode,ownership --force ${self'.packages.hyperspace-config-chain-2} $HYPERSPACE_DATA/config-chain-2.toml  
-                cp --dereference --no-preserve=mode,ownership --force ${self'.packages.hyperspace-config-chain-3} $HYPERSPACE_DATA/config-chain-3.toml  
-                cp --dereference --no-preserve=mode,ownership --force ${self'.packages.hyperspace-config-core} $HYPERSPACE_DATA/config-core.toml                
-                CODE_ID=$(cat /tmp/centauri-dev/code_id)
-                sed -i "s/wasm_code_id = \"0000000000000000000000000000000000000000000000000000000000000000\"/wasm_code_id = \"$CODE_ID\"/" "$HYPERSPACE_DATA/config-chain-2.toml"
-                ${self'.packages.hyperspace-composable-rococo-picasso-rococo}/bin/hyperspace create-clients --config-a $HYPERSPACE_DATA/config-chain-3.toml --config-b $HYPERSPACE_DATA/config-chain-2.toml --config-core $HYPERSPACE_DATA/config-core.toml --delay-period 10
-              '';
-              log_location = "/tmp/composable-devnet/hyperspace/clients.log";
-              depends_on = {
-                "centauri-init".condition = "process_completed_successfully";
-                "centauri".condition = "process_healthy";
-              };
-              availability = { restart = "on_failure"; };
-            };
-            hyperspace-connection = {
-              command = ''
-                COMPOSABLE_DATA=/tmp/composable-devnet/
-                HYPERSPACE_DATA="$COMPOSABLE_DATA/hyperspace"
-                RUST_LOG="hyperspace=trace,hyperspace_parachain=trace,hyperspace_cosmos=trace"
-                export RUST_LOG      
-                ${self'.packages.hyperspace-composable-rococo-picasso-rococo}/bin/hyperspace create-connection --config-a $HYPERSPACE_DATA/config-chain-3.toml --config-b $HYPERSPACE_DATA/config-chain-2.toml --config-core $HYPERSPACE_DATA/config-core.toml --delay-period 10
-              '';
-              log_location = "/tmp/composable-devnet/hyperspace/connection.log";
-              depends_on = {
-                "hyperspace-client".condition =
-                  "process_completed_successfully";
-              };
-              availability = { restart = "on_failure"; };
-            };
-            hyperspace-channels = {
-              command = ''
-                COMPOSABLE_DATA=/tmp/composable-devnet/
-                HYPERSPACE_DATA="$COMPOSABLE_DATA/hyperspace"
-                RUST_LOG="hyperspace=trace,hyperspace_parachain=trace,hyperspace_cosmos=trace"
-                export RUST_LOG
-                ${self'.packages.hyperspace-composable-rococo-picasso-rococo}/bin/hyperspace create-channel --config-a $HYPERSPACE_DATA/config-chain-3.toml --config-b $HYPERSPACE_DATA/config-chain-2.toml --config-core $HYPERSPACE_DATA/config-core.toml --delay-period 10 --port-id transfer --version ics20-1 --order unordered
-              '';
-              log_location = "/tmp/composable-devnet/hyperspace/channels.log";
-              depends_on = {
-                "hyperspace-connection".condition =
-                  "process_completed_successfully";
-              };
-              availability = { restart = "on_failure"; };
-            };
-            hyperspace-relay = {
-              command = ''
-                COMPOSABLE_DATA=/tmp/composable-devnet/
-                HYPERSPACE_DATA="$COMPOSABLE_DATA/hyperspace"
-                RUST_LOG="hyperspace=trace,hyperspace_parachain=trace,hyperspace_cosmos=trace"
-                export RUST_LOG
-                ${self'.packages.hyperspace-composable-rococo-picasso-rococo}/bin/hyperspace relay --config-a $HYPERSPACE_DATA/config-chain-3.toml --config-b $HYPERSPACE_DATA/config-chain-2.toml --config-core $HYPERSPACE_DATA/config-core.toml --delay-period 10
-              '';
-              log_location = "/tmp/composable-devnet/hyperspace/relay.log";
-              depends_on = {
-                "hyperspace-channels".condition =
-                  "process_completed_successfully";
-              };
-              availability = { restart = "on_failure"; };
-            };
+            #     cp --dereference --no-preserve=mode,ownership --force ${self'.packages.hyperspace-config-chain-2} $HYPERSPACE_DATA/config-chain-2.toml  
+            #     cp --dereference --no-preserve=mode,ownership --force ${self'.packages.hyperspace-config-chain-3} $HYPERSPACE_DATA/config-chain-3.toml  
+            #     cp --dereference --no-preserve=mode,ownership --force ${self'.packages.hyperspace-config-core} $HYPERSPACE_DATA/config-core.toml                
+            #     CODE_ID=$(cat /tmp/centauri-dev/code_id)
+            #     sed -i "s/wasm_code_id = \"0000000000000000000000000000000000000000000000000000000000000000\"/wasm_code_id = \"$CODE_ID\"/" "$HYPERSPACE_DATA/config-chain-2.toml"
+            #     ${self'.packages.hyperspace-composable-rococo-picasso-rococo}/bin/hyperspace create-clients --config-a $HYPERSPACE_DATA/config-chain-3.toml --config-b $HYPERSPACE_DATA/config-chain-2.toml --config-core $HYPERSPACE_DATA/config-core.toml --delay-period 10
+            #   '';
+            #   log_location = "/tmp/composable-devnet/hyperspace/clients.log";
+            #   depends_on = {
+            #     "centauri-init".condition = "process_completed_successfully";
+            #     "centauri".condition = "process_healthy";
+            #   };
+            #   availability = { restart = "on_failure"; };
+            # };
+            # hyperspace-connection = {
+            #   command = ''
+            #     COMPOSABLE_DATA=/tmp/composable-devnet/
+            #     HYPERSPACE_DATA="$COMPOSABLE_DATA/hyperspace"
+            #     RUST_LOG="hyperspace=trace,hyperspace_parachain=trace,hyperspace_cosmos=trace"
+            #     export RUST_LOG      
+            #     ${self'.packages.hyperspace-composable-rococo-picasso-rococo}/bin/hyperspace create-connection --config-a $HYPERSPACE_DATA/config-chain-3.toml --config-b $HYPERSPACE_DATA/config-chain-2.toml --config-core $HYPERSPACE_DATA/config-core.toml --delay-period 10
+            #   '';
+            #   log_location = "/tmp/composable-devnet/hyperspace/connection.log";
+            #   depends_on = {
+            #     "hyperspace-client".condition =
+            #       "process_completed_successfully";
+            #   };
+            #   availability = { restart = "on_failure"; };
+            # };
+            # hyperspace-channels = {
+            #   command = ''
+            #     COMPOSABLE_DATA=/tmp/composable-devnet/
+            #     HYPERSPACE_DATA="$COMPOSABLE_DATA/hyperspace"
+            #     RUST_LOG="hyperspace=trace,hyperspace_parachain=trace,hyperspace_cosmos=trace"
+            #     export RUST_LOG
+            #     ${self'.packages.hyperspace-composable-rococo-picasso-rococo}/bin/hyperspace create-channel --config-a $HYPERSPACE_DATA/config-chain-3.toml --config-b $HYPERSPACE_DATA/config-chain-2.toml --config-core $HYPERSPACE_DATA/config-core.toml --delay-period 10 --port-id transfer --version ics20-1 --order unordered
+            #   '';
+            #   log_location = "/tmp/composable-devnet/hyperspace/channels.log";
+            #   depends_on = {
+            #     "hyperspace-connection".condition =
+            #       "process_completed_successfully";
+            #   };
+            #   availability = { restart = "on_failure"; };
+            # };
+            # hyperspace-relay = {
+            #   command = ''
+            #     COMPOSABLE_DATA=/tmp/composable-devnet/
+            #     HYPERSPACE_DATA="$COMPOSABLE_DATA/hyperspace"
+            #     RUST_LOG="hyperspace=trace,hyperspace_parachain=trace,hyperspace_cosmos=trace"
+            #     export RUST_LOG
+            #     ${self'.packages.hyperspace-composable-rococo-picasso-rococo}/bin/hyperspace relay --config-a $HYPERSPACE_DATA/config-chain-3.toml --config-b $HYPERSPACE_DATA/config-chain-2.toml --config-core $HYPERSPACE_DATA/config-core.toml --delay-period 10
+            #   '';
+            #   log_location = "/tmp/composable-devnet/hyperspace/relay.log";
+            #   depends_on = {
+            #     "hyperspace-channels".condition =
+            #       "process_completed_successfully";
+            #   };
+            #   availability = { restart = "on_failure"; };
+            # };
           };
         };
       };
